Added -wall/pedantic

+++ src/Control/AnnounceStorage.c
@@ -51,7 +51,7 @@ AnnounceStorageHandle AnnounceStorage_create(const size_t maxAmountOfAddresses, 
     self->maxAmountOfChannels = maxAmountOfChannels;
 
     /* Initialize all addresses to have a defined state */
-    for(int i = 0; i < maxAmountOfAddresses; ++i){
+    for(size_t i = 0; i < maxAmountOfAddresses; ++i){
         self->addresses[i].type = SE_FLOAT;
         self->addresses[i].address = 0;
         self->addresses[i].name[0] = '\0';
